refactor: extract arbiter service functions from action.service.ts
Move markArgumentStrong and completeArbiterReview into their own
arbiter.service.ts to reduce action.service.ts from 1500+ lines
toward the 800-line limit.

Re-export both functions from action.service.ts so the controller's
wildcard import continues to work without changes. Update the test
file to import directly from arbiter.service.ts.
